// frontend/types/authentication/login.ts
import { z } from "zod";

export const loginSchema = z.object({
    username: z.string().min(1, "Username is required"),
    password: z.string().min(1, "Password is required"),
});

export type LoginFormData = z.infer<typeof loginSchema>;

export interface LoginResponse {
    accessToken: string;
    refreshToken: string;
}